Encapsulate 7.0 support into its own module
