+++ lib/active_record/connection_adapters/trilogy_adapter.rb
@@ -6,6 +6,8 @@ require "active_record/connection_adapters/abstract_mysql_adapter"
 require "active_record/tasks/trilogy_database_tasks"
 require "trilogy_adapter/lost_connection_exception_translator"
 
+require "trilogy_adapter/backwards_compatibility"
+
 module ActiveRecord
   module ConnectionAdapters
     class TrilogyAdapter < ::ActiveRecord::ConnectionAdapters::AbstractMysqlAdapter
@@ -155,53 +157,7 @@ module ActiveRecord
 
       # ActiveRecord 7.0 support
       if ActiveRecord.version < ::Gem::Version.new('7.1.a')
-        def initialize(connection, logger, connection_options, config)
-          super
-          if @connection
-            @raw_connection = @connection
-          end
-          # Ensure that we're treating prepared_statements in the same way that Rails 7.1 does
-          @prepared_statements = self.class.type_cast_config_to_boolean(
-            @config.fetch(:prepared_statements) { default_prepared_statements }
-          )
-        end
-
-        def connect!
-          verify!
-          self
-        end
-
-        def reconnect!
-          @lock.synchronize do
-            disconnect!
-            connect
-          rescue StandardError => original_exception
-            raise translate_exception_class(original_exception, nil, nil)
-          end
-        end
-
-        def with_trilogy_connection(uses_transaction: true, **_kwargs)
-          @lock.synchronize do
-            @raw_connection = @connection || nil unless instance_variable_defined?(:@raw_connection)
-            verify!
-            materialize_transactions if uses_transaction
-            yield @raw_connection
-          end
-        end
-
-        def execute(sql, name = nil, **kwargs)
-          sql = transform_query(sql)
-          check_if_write_query(sql)
-          super
-        end
-
-        def full_version
-          get_full_version
-        end
-
-        def default_timezone
-          ActiveRecord.default_timezone
-        end
+        include ::TrilogyAdapter::BackwardsCompatibility
       else # ActiveRecord >= 7.1 support
         alias_method :with_trilogy_connection, :with_raw_connection
       end
